Add destroyLoanScheme rpc
